+++ libs/sdk-core/CHANGELOG.md
@@ -1,14 +1,18 @@
 # 1.1.1 (unreleased)
 
 ### Improvements
+
 - Add pagination model in search results
+- Add `computedmetadata` to `IResource` model
 
 # 1.1.0 (2022-12-08)
 
 ### Breaking change
+
 - Remove widget endpoints: the widget is not stored in the backend anymore
 
 ### Improvements
+
 - Fix resource data if null
 - Add `getLabelFromFilter` and `getFilterFromLabel` functions
 - Add NUA key to the local storage only when the key is created from the desktop application